First draft implementation of organisations

+++ Modules/assessment/assessment_model.php
@@ -15,31 +15,52 @@ class Assessment
         $this->mysqli = $mysqli;
         $this->tablename = "assessment";
         $this->accesstable ="assessment_access";
+        
     }
     
-    public function admin($userid)
+    public function has_access($userid,$id)
     {
-        $userid = (int) $userid;
+        // Check if user has direct or shared access
+        $result = $this->mysqli->query("SELECT * FROM assessment_access WHERE `userid`='$userid' AND `id`='$id'");
+        if ($result->num_rows==1) return true;
         
-        // check if userid is an admin user or can only view own account
-        // $result = $this->mysqli->query("SELECT admin FROM ".$this->accesstable." WHERE `userid`='$userid'");
-        // $row = $result->fetch_array();
-        // if ($row && isset($row['admin']) && $row['admin']==1) return true;
+        $result = $this->mysqli->query("SELECT orgid FROM organisation_membership WHERE `userid`='$userid'");
+        while($row = $result->fetch_object()) {
+            $orgid = $row->orgid;
+            $result2 = $this->mysqli->query("SELECT * FROM assessment_access WHERE `orgid`='$orgid' AND `id`='$id'");
+            if ($result2->num_rows==1) return true;
+        }
+    }
+    
+    public function get_org_list($orgid)
+    {
+        $orgid = (int) $orgid;
+
+        $result = $this->mysqli->query("SELECT * FROM assessment_access WHERE `orgid`='$orgid'");
         
-        global $assessment_admin_users;
-        if (in_array($userid,$assessment_admin_users)) return true;
-        return false;
+        $projects = array();
+        while($row = $result->fetch_object()) {
+            $id = $row->id;
+            $assessment_result = $this->mysqli->query("SELECT id,name,description,userid,author,mdate,status FROM ".$this->tablename." WHERE `id`='$id'");
+            $row = $assessment_result->fetch_object();
+            $projects[] = $row;
+        }
+        return $projects;
     }
     
     public function get_list($userid)
     {
         $userid = (int) $userid;
 
-        $limit_to_user = ""; if (!$this->admin($userid)) $limit_to_user = " WHERE `userid`='$userid'";
-        $result = $this->mysqli->query("SELECT id,name,description,userid,author,mdate,status FROM ".$this->tablename.$limit_to_user);
+        $result = $this->mysqli->query("SELECT * FROM assessment_access WHERE `userid`='$userid'");
         
         $projects = array();
-        while($row = $result->fetch_object()) $projects[] = $row;
+        while($row = $result->fetch_object()) {
+            $id = $row->id;
+            $assessment_result = $this->mysqli->query("SELECT id,name,description,userid,author,mdate,status FROM ".$this->tablename." WHERE `id`='$id'");
+            $row = $assessment_result->fetch_object();
+            $projects[] = $row;
+        }
         return $projects;
     }
     
@@ -67,10 +88,13 @@ class Assessment
             'id'=>$id,
             'name'=>$name,
             'description'=>$description,
+            'status'=>$status,
             'userid'=>$userid,
             'author'=>$author,
             'mdate'=>$mdate
         );
+        
+        $this->access($id,$userid,1);
         
         return $project;
     }
@@ -79,11 +103,10 @@ class Assessment
     {
         $id = (int) $id;
         $userid = (int) $userid;
-        
-        $limit_to_user = ""; 
-        if (!$this->admin($userid)) $limit_to_user = " WHERE `userid`='$userid'";
+        if (!$this->has_access($userid,$id)) return false;
         
         $result = $this->mysqli->query("DELETE FROM ".$this->tablename." WHERE `id` = '$id'");
+        $result = $this->mysqli->query("DELETE FROM assessment_access WHERE `id` = '$id'");
         return array("Deleted");
     }
     
@@ -91,10 +114,8 @@ class Assessment
     {
         $id = (int) $id;
         $userid = (int) $userid;
-        
-        $limit_to_user = ""; 
-        if (!$this->admin($userid)) $limit_to_user = " WHERE `userid`='$userid'";
-        
+        if (!$this->has_access($userid,$id)) return false;
+            
         $result = $this->mysqli->query("SELECT * FROM ".$this->tablename." WHERE `id` = '$id'");
         $row = $result->fetch_object();
         
@@ -107,11 +128,9 @@ class Assessment
         $id = (int) $id;
         $userid = (int) $userid;
         $status = preg_replace('/[^\w\s]/','',$status);
+        if (!$this->has_access($userid,$id)) return false;
         
-        $limit_to_user = ""; 
-        if (!$this->admin($userid)) $limit_to_user = " AND `userid`='$userid'";
-        print $id;
-        $stmt = $this->mysqli->prepare("UPDATE ".$this->tablename." SET `status` = ? WHERE `id` = ?".$limit_to_user);
+        $stmt = $this->mysqli->prepare("UPDATE ".$this->tablename." SET `status` = ? WHERE `id` = ?");
         $stmt->bind_param("si", $status, $id);
         $stmt->execute();
         
@@ -122,9 +141,7 @@ class Assessment
     {
         $id = (int) $id;
         $userid = (int) $userid;
-        
-        $limit_to_user = ""; 
-        if (!$this->admin($userid)) $limit_to_user = " AND `userid`='$userid'";
+        if (!$this->has_access($userid,$id)) return false;
         
         $data = preg_replace('/[^\w\s-.",:{}\[\]]/','',$data);
         $data = json_decode($data);
@@ -136,7 +153,7 @@ class Assessment
 
             $data = json_encode($data);
           
-            $stmt = $this->mysqli->prepare("UPDATE ".$this->tablename." SET `data` = ?, `mdate` = ? WHERE `id` = ?".$limit_to_user);
+            $stmt = $this->mysqli->prepare("UPDATE ".$this->tablename." SET `data` = ?, `mdate` = ? WHERE `id` = ?");
             $stmt->bind_param("ssi", $data, $mdate, $id);
             $stmt->execute();
 
@@ -145,5 +162,50 @@ class Assessment
         else {
           return false;
         }
+    }
+    
+    public function access($id,$userid,$write)
+    {
+        $result = $this->mysqli->query("INSERT INTO assessment_access SET `id` = '$id', `userid` = '$userid', `orgid` = '0', `write` = '$write'");
+    }
+    
+    public function org_access($id,$orgid,$write)
+    {
+        $result = $this->mysqli->query("INSERT INTO assessment_access SET `id` = '$id', `userid` = '0', `orgid` = '$orgid', `write` = '$write'");
+    }
+    
+    public function share($id,$username)
+    {
+        global $user;
+        
+        // 1. Check if user exists
+        $userid = $user->get_id($username);
+        if ($userid===false) return "User does not exist";
+        
+        // 2. Check if already shared with user
+        $result = $this->mysqli->query("SELECT * FROM assessment_access WHERE `id` = '$id' AND `userid`='$userid'");
+        if ($result->num_rows==1) return "Already shared";
+        
+        // 3. Register share
+        $this->access($id,$userid,1);
+        return "Assessment shared";
+    }
+    
+    public function getshared($id)
+    {
+        $result = $this->mysqli->query("SELECT * FROM assessment_access WHERE `id` = '$id'");
+        $users = array();
+        while($row = $result->fetch_object()) {
+            global $user;
+            if ($row->userid!=0) $username = $user->get_name($row->userid);
+            if ($row->orgid!=0) {
+                $orgid = $row->orgid;
+                $orgresult = $this->mysqli->query("SELECT * FROM organisations WHERE `id`='$orgid'");
+                $orgrow = $orgresult->fetch_object();
+                $username = $orgrow->name;
+            }
+            $users[] = array('orgid'=>$row->orgid, 'userid'=>$row->userid, 'username'=>$username);
+        }
+        return $users;
     }
 }
